web ui/requirements.txt
Flask==2.1.2
requests==2.26.0
python-dotenv
Werkzeug